requirements.txt
mysql-connector-python==8.3.0
python-dotenv==1.0.1
google-generativeai==0.5.2
langgraph==0.0.35
